feat: add vite as bundler, fix minor bugs
