+++ index.html
@@ -3,7 +3,7 @@
 
 <head>
   <meta charset="utf-8" />
-  <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
+  <link rel="icon" href="/favicon.ico" />
   <link rel="stylesheet" href="https://use.typekit.net/ure3vlz.css">
   <link rel="preconnect" href="https://fonts.googleapis.com">
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
@@ -12,12 +12,12 @@
   <meta name="theme-color" content="#000000" />
   <meta name="description" content="Discover Immortal by Deeli.ai, the AI-driven SaaS for tech scouting and trend analysis. Empowering VCs and innovation centers to identify early-stage tech and startups." />
   <meta name="google-site-verification" content="MOZJXFv_YO2qJpJkd5f_jZw2moAUp15i_GRcziW1-uA" />
-  <link rel="apple-touch-icon" href="%PUBLIC_URL%/favicon.ico" />
+  <link rel="apple-touch-icon" href="/favicon.ico" />
   
   <!-- Open Graph Meta Tags -->
   <meta property="og:title" content="Revolutionize Your Tech Scouting and Intelligence with AI | Deeli.ai" />
   <meta property="og:description" content="Discover Immortal by Deeli.ai, the AI-driven SaaS for tech scouting and trend analysis. Empowering VCs and innovation centers to identify early-stage tech and startups." />
-  <meta property="og:image" content="%PUBLIC_URL%/main.png" />
+  <meta property="og:image" content="/main.png" />
   <meta property="og:url" content="https://deeli.ai" />
   <meta property="og:type" content="website" />
   
@@ -26,9 +26,9 @@
   <meta name="twitter:title" content="Deeli AI" />
   <meta name="twitter:description" content="Discover Immortal by Deeli.ai, the AI-driven SaaS for tech scouting and trend analysis. Empowering VCs and innovation centers to identify early-stage tech and startups.
 " />
-  <meta name="twitter:image" content="%PUBLIC_URL%/main.png" />
+  <meta name="twitter:image" content="/main.png" />
   
-  <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+  <link rel="manifest" href="/manifest.json" />
   
   <title>Deeli AI</title>
 </head>
@@ -37,5 +37,5 @@
   <noscript>You need to enable JavaScript to run this app.</noscript>
   <div id="root"></div>
 </body>
-
+<script type="module" src="/src/index.jsx"></script>
 </html>
