+++ package.json
@@ -2,6 +2,7 @@
   "name": "company-website",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "dependencies": {
     "@dotlottie/player-component": "^2.7.12",
     "@emotion/react": "^11.11.4",
@@ -28,7 +29,6 @@
     "react-intersection-observer": "^9.10.3",
     "react-lottie": "^1.2.4",
     "react-lottie-player": "^2.0.0",
-    "react-scripts": "5.0.1",
     "react-scroll": "^1.9.0",
     "react-slick": "^0.30.2",
     "react-spring": "^9.7.3",
@@ -36,12 +36,13 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
+    "dev": "vite serve",
+    "start": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "test": "vite test"
   },
   "eslintConfig": {
     "extends": [
@@ -62,6 +63,8 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^6.1.1"
+    "@vitejs/plugin-react": "^4.3.1",
+    "gh-pages": "^6.1.1",
+    "vite": "^5.4.2"
   }
 }
