+++ src/components/LottieAnimation.jsx
