+++ src/components/Section3.jsx
@@ -42,6 +42,7 @@ const Section3 = () => {
   const [ref2, inView2] = useInView({ triggerOnce: true, threshold: 0.15 });
   const [ref3, inView3] = useInView({ triggerOnce: true, threshold: 0.15 });
   const sectionRef = useSectionTracker("section3");
+  
   useEffect(() => {
     if (inView1 || inView2 || inView3) {
       setStartAnimation(true);
